Updated the unknown items to TP7 and TP8 Data in the Packet Layout. These were determined by looking at the extra sensors available and cross-referencing them to the 10-20 International Electrode Placement diagram. Those locations were then cross-referenced by Emotiv Headset Setup guide images. Future tests could confirm this via selective channel data gathering.
In addition switched the AF3 Data -> AF7 Data and AF4 Data -> AF8 in the packet layout. The electrode placement did not seem to line up on scalp for most users to correlate with areas AF3/AF4. In addition when referencing the Emotiv Headset Setup Guide with the 10-20 International Electrode Placement Diagram that also confirmed that change. 

-Emily O'Leary 07.19.2012